Support c-ares, Refactor DNS (#4275)
* add core tests

* fix

* optimize

* optimize [2]

* optimize [3]

* optimize [4]

* optimize [5]

* Refactor

* Refactor [2]

* Refactor [3]

* fix

* fix core tests

* fix tests

* fix core tests

+++ ext-src/swoole_mysql_coro.cc
@@ -927,11 +927,11 @@ void mysql_client::handle_row_data_text(zval *return_value, mysql::row_data *row
         RETVAL_STRINGL(p, row_data->text.length);
     _return:
         swTraceLog(SW_TRACE_MYSQL_CLIENT,
-                   "%.*s=[%zu]%.*s%s",
+                   "%.*s=[%lu]%.*s%s",
                    field->name_length,
                    field->name,
                    Z_STRLEN_P(return_value),
-                   SW_MIN(32, Z_STRLEN_P(return_value)),
+                   (int) SW_MIN(32, Z_STRLEN_P(return_value)),
                    Z_STRVAL_P(return_value),
                    (Z_STRLEN_P(return_value) > 32 ? "..." : ""));
     }
@@ -1247,36 +1247,36 @@ void mysql_statement::send_execute_request(zval *return_value, zval *params) {
         zval *value;
         ZEND_HASH_FOREACH_VAL(Z_ARRVAL_P(params), value) {
             switch (client->strict_type ? Z_TYPE_P(value) : (IS_NULL == Z_TYPE_P(value) ? IS_NULL : IS_STRING)) {
-                case IS_NULL:
-                    *((buffer->str + null_start_offset) + (index / 8)) |= (1UL << (index % 8));
-                    sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_NULL);
-                    break;
-                case IS_TRUE:
-                case IS_FALSE:
-                case IS_LONG:
-                    sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_LONGLONG);
-                    sw_mysql_int8store(stack_buffer, zval_get_long(value));
-                    if (buffer->append(stack_buffer, mysql::get_static_type_size(SW_MYSQL_TYPE_LONGLONG)) < 0) {
-                        RETURN_FALSE;
-                    }
-                    break;
-                case IS_DOUBLE:
-                    sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_DOUBLE);
-                    sw_mysql_doublestore(stack_buffer, zval_get_double(value));
-                    if (buffer->append(stack_buffer, mysql::get_static_type_size(SW_MYSQL_TYPE_DOUBLE)) < 0) {
-                        RETURN_FALSE;
-                    }
-                    break;
-                default:
-                    zend::String str_value(value);
-                    uint8_t lcb_size = mysql::write_lcb(stack_buffer, str_value.len());
-                    sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_VAR_STRING);
-                    if (buffer->append(stack_buffer, lcb_size) < 0) {
-                        RETURN_FALSE;
-                    }
-                    if (buffer->append(str_value.val(), str_value.len()) < 0) {
-                        RETURN_FALSE;
-                    }
+            case IS_NULL:
+                *((buffer->str + null_start_offset) + (index / 8)) |= (1UL << (index % 8));
+                sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_NULL);
+                break;
+            case IS_TRUE:
+            case IS_FALSE:
+            case IS_LONG:
+                sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_LONGLONG);
+                sw_mysql_int8store(stack_buffer, zval_get_long(value));
+                if (buffer->append(stack_buffer, mysql::get_static_type_size(SW_MYSQL_TYPE_LONGLONG)) < 0) {
+                    RETURN_FALSE;
+                }
+                break;
+            case IS_DOUBLE:
+                sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_DOUBLE);
+                sw_mysql_doublestore(stack_buffer, zval_get_double(value));
+                if (buffer->append(stack_buffer, mysql::get_static_type_size(SW_MYSQL_TYPE_DOUBLE)) < 0) {
+                    RETURN_FALSE;
+                }
+                break;
+            default:
+                zend::String str_value(value);
+                uint8_t lcb_size = mysql::write_lcb(stack_buffer, str_value.len());
+                sw_mysql_int2store((buffer->str + type_start_offset) + (index * 2), SW_MYSQL_TYPE_VAR_STRING);
+                if (buffer->append(stack_buffer, lcb_size) < 0) {
+                    RETURN_FALSE;
+                }
+                if (buffer->append(str_value.val(), str_value.len()) < 0) {
+                    RETURN_FALSE;
+                }
             }
             index++;
         }
@@ -1483,11 +1483,10 @@ void mysql_statement::fetch(zval *return_value) {
                 case SW_MYSQL_TYPE_LONGLONG:
                     if (field->flags & SW_MYSQL_UNSIGNED_FLAG) {
                         add_assoc_ulong_safe_ex(return_value, field->name, field->name_length, *(uint64_t *) p);
-                        swTraceLog(
-                            SW_TRACE_MYSQL_CLIENT, "%.*s=%llu", field->name_length, field->name, *(uint64_t *) p);
+                        swTraceLog(SW_TRACE_MYSQL_CLIENT, "%.*s=%lu", field->name_length, field->name, *(uint64_t *) p);
                     } else {
                         add_assoc_long_ex(return_value, field->name, field->name_length, *(int64_t *) p);
-                        swTraceLog(SW_TRACE_MYSQL_CLIENT, "%.*s=%lld", field->name_length, field->name, *(int64_t *) p);
+                        swTraceLog(SW_TRACE_MYSQL_CLIENT, "%.*s=%ld", field->name_length, field->name, *(int64_t *) p);
                     }
                     break;
                 case SW_MYSQL_TYPE_FLOAT: {
@@ -1901,7 +1900,8 @@ static PHP_METHOD(swoole_mysql_coro, fetch) {
     mc->fetch(return_value);
     mc->del_timeout_controller();
     if (sw_unlikely(Z_TYPE_P(return_value) == IS_FALSE)) {
-        swoole_mysql_coro_sync_error_properties(ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
+        swoole_mysql_coro_sync_error_properties(
+            ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
     }
 }
 
@@ -1918,7 +1918,8 @@ static PHP_METHOD(swoole_mysql_coro, fetchAll) {
     mc->fetch_all(return_value);
     mc->del_timeout_controller();
     if (sw_unlikely(Z_TYPE_P(return_value) == IS_FALSE)) {
-        swoole_mysql_coro_sync_error_properties(ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
+        swoole_mysql_coro_sync_error_properties(
+            ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
     }
 }
 
@@ -1958,7 +1959,8 @@ static PHP_METHOD(swoole_mysql_coro, prepare) {
     mc->add_timeout_controller(timeout, Socket::TIMEOUT_RDWR);
     if (UNEXPECTED(!mc->send_prepare_request(statement, statement_length))) {
     _failed:
-        swoole_mysql_coro_sync_error_properties(ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
+        swoole_mysql_coro_sync_error_properties(
+            ZEND_THIS, mc->get_error_code(), mc->get_error_msg(), mc->is_connected());
         RETVAL_FALSE;
     } else if (UNEXPECTED(mc->get_defer())) {
         RETVAL_TRUE;
